added better debug info in RsGenexchange (allowing service-based debug) and fixed growing of mGrpNotify and mMsgNotify in DataAccess due to not acknowledging tokens for explicit operations in circles

+++ libretroshare/src/services/p3gxscircles.cc
@@ -601,7 +601,7 @@ void p3GxsCircles::notifyChanges(std::vector<RsGxsNotify *> &changes)
 #endif
 			RsGxsCircleId circle_id(msgChange->mGroupId);
 
-			if(rsEvents && (c->getType() == RsGxsNotify::TYPE_RECEIVED_NEW))
+            if(rsEvents && ((c->getType() == RsGxsNotify::TYPE_RECEIVED_NEW) || (c->getType() == RsGxsNotify::TYPE_PUBLISHED)))
 			{
 				const RsGxsCircleSubscriptionRequestItem *item = dynamic_cast<const RsGxsCircleSubscriptionRequestItem *>(msgChange->mNewMsgItem);
 
@@ -807,7 +807,7 @@ bool p3GxsCircles::getCircleDetails(const RsGxsCircleId& id, RsGxsCircleDetails&
 			details.mRestrictedCircleId = data.mRestrictedCircleId;
 
 			details.mAllowedNodes = data.mAllowedNodes;
-			details.mSubscriptionFlags.clear();
+            details.mSubscriptionFlags.clear();
 			details.mAllowedGxsIds.clear();
 			details.mAmIAllowed = false ;
 			details.mAmIAdmin = bool(data.mGroupSubscribeFlags & GXS_SERV::GROUP_SUBSCRIBE_ADMIN);
@@ -1501,6 +1501,72 @@ bool p3GxsCircles::checkCircleCache()
 	return true ;
 }
 
+bool p3GxsCircles::locked_setGroupUnprocessedStatus(RsGxsCircleCache& cache,bool unprocessed)
+{
+    uint32_t token2;
+
+    if(unprocessed)
+        cache.mGroupStatus |=  GXS_SERV::GXS_GRP_STATUS_UNPROCESSED;
+    else
+        cache.mGroupStatus &= ~GXS_SERV::GXS_GRP_STATUS_UNPROCESSED;
+
+    RsGenExchange::setGroupStatusFlags(token2, RsGxsGroupId(cache.mCircleId), unprocessed, GXS_SERV::GXS_GRP_STATUS_UNPROCESSED);
+
+    std::cerr << "********** new token for setGrpStatusFlags: " << token2 << std::endl;
+
+    // Now we need to async acknowledge the token when the job is finished. We cannot do this sync because it's the
+    // current thread that takes care of calling the handling of group processing.
+
+    RsThread::async([token2,this]()
+    {
+        std::chrono::milliseconds maxWait = std::chrono::milliseconds(10000);
+        std::chrono::milliseconds checkEvery = std::chrono::milliseconds(100);
+
+        auto timeout = std::chrono::steady_clock::now() + maxWait;	// wait for 10 secs at most
+        auto st = requestStatus(token2);
+
+        while( !(st == RsTokenService::FAILED || st >= RsTokenService::COMPLETE) && std::chrono::steady_clock::now() < timeout )
+        {
+            std::this_thread::sleep_for(checkEvery);
+            st = requestStatus(token2);
+        }
+
+        RsGxsGroupId grpId;
+        acknowledgeGrp(token2,grpId);
+    });
+    return true;
+}
+
+bool p3GxsCircles::locked_subscribeToCircle(const RsGxsCircleId &grpId, bool subscribe)
+{
+    uint32_t token;
+    if(!RsGenExchange::subscribeToGroup(token, RsGxsGroupId(grpId), subscribe))
+        return false;
+
+    // Now we need to async acknowledge the token when the job is finished. We cannot do this sync because it's the
+    // current thread that takes care of calling the handling of group processing.
+
+    RsThread::async([token,this]()
+    {
+        std::chrono::milliseconds maxWait = std::chrono::milliseconds(10000);
+        std::chrono::milliseconds checkEvery = std::chrono::milliseconds(100);
+
+        auto timeout = std::chrono::steady_clock::now() + maxWait;	// wait for 10 secs at most
+        auto st = requestStatus(token);
+
+        while( !(st == RsTokenService::FAILED || st >= RsTokenService::COMPLETE) && std::chrono::steady_clock::now() < timeout )
+        {
+            std::this_thread::sleep_for(checkEvery);
+            st = requestStatus(token);
+        }
+
+        RsGxsGroupId grpId;
+        acknowledgeGrp(token,grpId);
+    });
+
+    return true;
+}
+
 bool p3GxsCircles::locked_checkCircleCacheForMembershipUpdate(RsGxsCircleCache& cache)
 {
 	rstime_t now = time(NULL) ;
@@ -1513,20 +1579,19 @@ bool p3GxsCircles::locked_checkCircleCacheForMembershipUpdate(RsGxsCircleCache& 
 #ifdef DEBUG_CIRCLES
 		std::cerr << "Cache entry for circle " << cache.mCircleId << " needs a swab over membership requests. Re-scheduling it." << std::endl;
 #endif
-        cache.mGroupStatus |= GXS_SERV::GXS_GRP_STATUS_UNPROCESSED;	// forces processing of cache entry
-        uint32_t token;
-        RsGenExchange::setGroupStatusFlags(token, RsGxsGroupId(cache.mCircleId.toStdString()), 0, GXS_SERV::GXS_GRP_STATUS_UNPROCESSED);
+        locked_setGroupUnprocessedStatus(cache,true); // forces the re-check of the group
 
-		// this should be called regularly
+        // this should be called regularly
 
         RsTokReqOptions opts;
 		opts.mReqType = GXS_REQUEST_TYPE_MSG_DATA;
 		std::list<RsGxsGroupId> grpIds ;
+        uint32_t token2;
 
 		grpIds.push_back(RsGxsGroupId(cache.mCircleId)) ;
 
-		RsGenExchange::getTokenService()->requestMsgInfo(token, RS_TOKREQ_ANSTYPE_SUMMARY,	opts, grpIds);
-		GxsTokenQueue::queueRequest(token, CIRCLEREQ_MESSAGE_DATA);	
+        RsGenExchange::getTokenService()->requestMsgInfo(token2, RS_TOKREQ_ANSTYPE_SUMMARY,	opts, grpIds);
+        GxsTokenQueue::queueRequest(token2, CIRCLEREQ_MESSAGE_DATA);
 	}
 	return true ;
 }
@@ -1599,8 +1664,7 @@ bool p3GxsCircles::locked_checkCircleCacheForAutoSubscribe(RsGxsCircleCache& cac
             /* we are part of this group - subscribe, clear unprocessed flag */
             std::cerr << "  either admin or have posted a subscribe/unsubscribe message => AutoSubscribing!" << std::endl;
 #endif
-            uint32_t token;
-            RsGenExchange::subscribeToGroup(token, RsGxsGroupId(cache.mCircleId), true);
+            locked_subscribeToCircle(cache.mCircleId,true);
             mShouldSendCacheUpdateNotification = true;
         }
 #ifdef DEBUG_CIRCLES
@@ -1617,8 +1681,7 @@ bool p3GxsCircles::locked_checkCircleCacheForAutoSubscribe(RsGxsCircleCache& cac
         /* we know all the peers - we are not part - we can flag as PROCESSED. */
 		if(cache.mGroupSubscribeFlags & GXS_SERV::GROUP_SUBSCRIBE_SUBSCRIBED)
 		{
-            uint32_t token;
-            RsGenExchange::subscribeToGroup(token, RsGxsGroupId(cache.mCircleId), false);
+            locked_subscribeToCircle(cache.mCircleId,false);
             mShouldSendCacheUpdateNotification = true;
 #ifdef DEBUG_CIRCLES
             std::cerr << "  Neither admin nor subscription msg author! Let's unsubscribe this circle of unfriendly Napoleons!" << std::endl;
@@ -1634,9 +1697,7 @@ bool p3GxsCircles::locked_checkCircleCacheForAutoSubscribe(RsGxsCircleCache& cac
 #ifdef DEBUG_CIRCLES
     std::cerr << "  Marking the cache entry as processed." << std::endl;
 #endif
-    uint32_t token2;
-    cache.mGroupStatus &= ~GXS_SERV::GXS_GRP_STATUS_UNPROCESSED;
-    RsGenExchange::setGroupStatusFlags(token2, RsGxsGroupId(cache.mCircleId), 0, GXS_SERV::GXS_GRP_STATUS_UNPROCESSED);
+    locked_setGroupUnprocessedStatus(cache,false);
 
     return true;
 }
@@ -1830,9 +1891,7 @@ void p3GxsCircles::handle_event(uint32_t event_type, const std::string &elabel)
 //                    |             |   Grp Subscribed: NO         |   Grp Subscribed: NO        |         
 //                    +-------------+------------------------------+-----------------------------+
 
-bool p3GxsCircles::pushCircleMembershipRequest(
-        const RsGxsId& own_gxsid, const RsGxsCircleId& circle_id,
-        RsGxsCircleSubscriptionType request_type )
+bool p3GxsCircles::pushCircleMembershipRequest( const RsGxsId& own_gxsid, const RsGxsCircleId& circle_id, RsGxsCircleSubscriptionType request_type )
 {
 	Dbg3() << __PRETTY_FUNCTION__ << "own_gxsid = " << own_gxsid
 	       << ", circle=" << circle_id << ", req type=" << request_type
@@ -1859,10 +1918,7 @@ bool p3GxsCircles::pushCircleMembershipRequest(
     // If the circle is not subscribed, then subscribe, whatever the subscription type. Indeed, if we publish a msg, even a msg for
     // unsubscribing, we need to have a subscribed group first.
 
-    uint32_t token ;
-    RsGenExchange::subscribeToGroup(token, RsGxsGroupId(circle_id), true);
-
-    if(waitToken(token) != RsTokenService::COMPLETE)
+    if(!locked_subscribeToCircle(circle_id,true))
     {
         std::cerr << __PRETTY_FUNCTION__ << " Could not subscribe to Circle group." << std::endl;
         return false;
@@ -1897,7 +1953,29 @@ bool p3GxsCircles::pushCircleMembershipRequest(
     std::cerr << "  ThreadId   : " << s->meta.mThreadId << std::endl;
 #endif
 
+    uint32_t token;
     RsGenExchange::publishMsg(token, s);
+
+    // This is manual handling of token. We need to clear it up from the notification when done, and that needs
+    // to be async-ed, since the processing of message publication is done in the same thread.
+
+    RsThread::async( [this,token]()
+    {
+        std::chrono::milliseconds maxWait = std::chrono::milliseconds(10000);
+        std::chrono::milliseconds checkEvery = std::chrono::milliseconds(100);
+
+        auto timeout = std::chrono::steady_clock::now() + maxWait;	// wait for 10 secs at most
+        auto st = requestStatus(token);
+
+        while( !(st == RsTokenService::FAILED || st >= RsTokenService::COMPLETE) && std::chrono::steady_clock::now() < timeout )
+        {
+            std::this_thread::sleep_for(checkEvery);
+            st = requestStatus(token);
+        }
+
+        std::pair<RsGxsGroupId,RsGxsMessageId> grpmsgId;
+        acknowledgeMsg(token,grpmsgId);
+    });
     
     // update the cache.
     force_cache_reload(circle_id);
@@ -2046,9 +2124,12 @@ bool p3GxsCircles::processMembershipRequests(uint32_t token)
         locked_checkCircleCacheForAutoSubscribe(cache);
     }
 
-    RsStackMutex stack(mCircleMtx); /********** STACK LOCKED MTX ******/
-    uint32_t token2;
-    RsGenExchange::deleteMsgs(token2,messages_to_delete);
+    if(!messages_to_delete.empty())
+    {
+        RsStackMutex stack(mCircleMtx); /********** STACK LOCKED MTX ******/
+        uint32_t token2;
+        RsGenExchange::deleteMsgs(token2,messages_to_delete);
+    }
     return true ;
 }
